+++ lib/main_dev.dart
@@ -1,0 +1,9 @@
+import 'package:craft_flutter_template/app/bootstrap.dart';
+import 'package:craft_flutter_template/core/core.dart';
+
+/// Development entry point.
+///
+/// ```sh
+/// flutter run --flavor dev --target lib/main_dev.dart
+/// ```
+void main() => bootstrap(AppEnv.development);